vote_app_api/voteproject/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
Django==4.2
django-cors-headers==4.3.1
django-filter==23.5
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.3
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
pillow==10.2.0
pycparser==2.21
pyotp==2.9.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
